Struct cargonauts::clients::Conn

pub struct Conn<C> where
    C: 'static + NewService,  { /* fields omitted */ }

A smart wrapper around a connection which stores it back in the pool when it is dropped.

This can be dereferences to the Service instance this pool manages, and also implements Service itself by delegating.

Methods

impl<C> Conn<C> where
C: 'static + NewService,

fn new_unpooled(instance: <C as NewService>::Instance) -> Conn<C>

This constructor creates a connection which is not stored in a thread pool. It can be useful for purposes in which you need to treat a non-pooled connection as if it were stored in a pool, such as during tests.