Module cargonauts::middleware::http

Structs

Request
A request bundles several parts of an incoming NetworkStream, given to a Handler.

Response
The Response sent to a client after receiving a Request in a Service.

Enums

Error
A set of errors that can occur parsing HTTP streams.

Type Definitions

BoxFuture