Struct cargonauts::server::pool::Pool

pub struct Pool<C> where C: NewService { /* fields omitted */ }

An asynchronous, single-threaded connection pool.

This pool stores connections for re-use according to the policies set by the Config type. It uses an asynchronous tokio event loop, and performs the connections asynchronously. It can take any type of connection which implements a tokio protocol.

The first type parameter is the protocol for the clients produced by this pool. The second parameter is the Kind type, usually found in tokio_proto, which is used to distinguish pipelined and mutiplexed connections.

Methods

impl<C> Pool<C> where C: 'static + NewService

fn new(client: C,
handle: Handle,
config: Config)
-> Box<Future<Error = Error, Item = Pool<C>> + 'static>

Construct a new pool. This returns a future, because it will attempt to establish the minimum number of connections immediately.

This takes an address and a protocol for establishing connections, a handle to an event loop to run those connections on, and a configuration object to control its policy.

fn connection<E>(&self)
-> Either<FutureResult<Conn<C>, E>, Box<Future<Error = E, Item = Conn<C>> + 'static>> where E: From<Error>

Yield a connection from the pool.

In the happy path, this future will evaluate immediately and perform no allocations - it just pulls an idle connection out of the pool.

In the less happy path, depending on the state of the pool and your configurations, it may do one of several things:

It may attempt to establish a new connection

It may wait for a connection to become available and use that one.

It may or may not timeout while waiting, depending on your configuration.

Once the connection this future yields is dropped, it will be returned the pool. During storage, the connection may be released according to your configuration. Otherwise, it will prioritize giving the connection to a waiting request and only if there are none return it to the queue inside the pool.