Struct cargonauts::server::Request

pub struct Request { /* fields omitted */ }

A request bundles several parts of an incoming NetworkStream, given to a Handler.

Methods

impl Request

fn method(&self) -> &Method

The Method, such as Get, Post, etc.

fn headers(&self) -> &Headers

The headers of the incoming request.

fn uri(&self) -> &Uri

The target request-uri for this request.

fn version(&self) -> HttpVersion

The version of HTTP for this request.

fn remote_addr(&self) -> Option<SocketAddr>

The remote socket address of this request

This is an Option, because some underlying transports may not have a socket address, such as Unix Sockets.

fn path(&self) -> &str

The target path of this Request.

fn query(&self) -> Option<&str>

The query string of this Request.

fn body(self) -> Body

Take the Body of this Request.

fn deconstruct(self) -> (Method, Uri, HttpVersion, Headers, Body)

Deconstruct this Request into its pieces.

Modifying these pieces will have no effect on how hyper behaves.