Struct cargonauts::Environment

pub struct Environment { /* fields omitted */ }

The Environment in which an API endpoint runs.

The resource method, format, and middlware of an endpoint all have access to this environment. It is used for a few basic purposes: * It controls access to the connections to other network services. * It allows you to cache values in a type map to share them from the format/middleware/resource.

Methods

impl Environment

fn store<T>(&self, val: T) -> Option<T> where
T: Any,

Store a new value in the environment.

If a value of this type was already stored in the environment, that previous value is returned.

fn take<T>(&self) -> Option<T> where
T: Any,

Take a value of a type from the environment.

fn contains<T>(&self) -> bool where
T: Any,

Check if the environment contains a value of the given type.

fn conn<C>(
&self
) -> Either<FutureResult<Conn<C>, Error>, Box<Future<Error = Error, Item = Conn<C>> + 'static>> where
C: NewService,

Check out a connection to a service from the connection pool.

fn client<C>(&self) -> Box<Future<Error = Error, Item = C> + 'static> where
C: ConnectClient<<C as Client>::Connection>,

Check out a high level client from the environment.